Reference NUMBER will from now on be mapped to BigDecimal instead of Float
--HG--
extra : rebase_source : 49c4179f8cbe7e6810744e8cc9073c788637e2d6

+++ src/org/openbravo/base/model/Reference.java
@@ -66,7 +66,7 @@ public class Reference extends ModelObject {
     primitiveTypes.put("16", Date.class);
     primitiveTypes.put("17", String.class);
     primitiveTypes.put("20", Boolean.class);
-    primitiveTypes.put("22", Float.class);
+    primitiveTypes.put("22", BigDecimal.class);
     primitiveTypes.put("23", byte[].class); // Binary/Blob Data
     primitiveTypes.put("24", Timestamp.class);
     primitiveTypes.put("26", Object.class); // RowID is not used
